Encapsulate prison data in a model
This provides a class that wraps a prisons data as set in the prison config file, exposing methods to query it.

It exposes a class method for performing lookups based on a Prisons name.

The interface for an instance of the said model currently only exposes methods that are required for schedule related queries(e.g checking if a given day is one in which a prison is expected to be working).