Use Time.zone instead of Date.today
Make use of the apps configured `London` timezone when 
getting todays date.

Using Date.today will use the time on the server and not
the time zone configured by the application. This can lead
some indetermination when running code outside of a
developers workspace. We spotted this when specs failed on
CircleCI.

For more info about timezones and Rails [this article](http://danilenko.org/2012/7/6/rails_timezones/)
is worth a read.